Remove duplicated code for building remote server from source
